backend/requirements.txt
strands-agents>=0.1.0
flask>=3.0.0
flask-cors>=4.0.0
boto3>=1.34.0
mcp>=1.0.0
